refactor: `aside-icons` → `aside-utils`
