+++ packages/starlight/user-components/Aside.astro
@@ -2,7 +2,7 @@
 import { AstroError } from 'astro/errors';
 import Icon from './Icon.astro';
 import type { StarlightIcon } from '../components-internals/Icons';
-import { asideVariants, getAsideIconName, type AsideVariant } from '../integrations/aside-icons';
+import { asideVariants, getAsideIconName, type AsideVariant } from '../integrations/aside-utils';
 
 interface Props {
 	type?: AsideVariant;
